OBC/IC/forcing run BLING

+++ pre-processing/forcing/iron_dust/create_forcing_iron_dust_ASTE.py
@@ -19,7 +19,7 @@ aste_iron = create_ic()
 
 for var in list_bgc_variables:
 	for kt in np.arange(12):
-		aste_iron.create_ic_regional_llc(var,list_dataset,'iron_month' + str(kt+1).zfill(2),precision='single',timestep=kt)
+		aste_iron.create_ic_regional_llc(var,list_dataset,'iron_month' + str(kt+1).zfill(2),precision='double',timestep=kt)
 
 # concat all the files
 fileout = dir_ic_out + 'iron_deposition.bin'
